Fix Kotlin compile errors (array compound assignment, Double LUB in ln)
- RealFFT: use explicit re[i]=re[i]/n (DoubleArray /= not accepted).
- VocosFrontend: force MEL_MIN branch to Double so ln(clamped) typechecks.
- VocosVocoder/Synthesizer: explicit array element updates (defensive).

+++ app/src/main/java/com/dhvaani/app/dsp/VocosFrontend.kt
@@ -90,7 +90,7 @@ class VocosFrontend(
             for (f in 0 until nFreq) {
                 acc += magnitude[f].toDouble() * melFilterbank[base + f].toDouble()
             }
-            val clamped = if (acc < DspConstants.MEL_MIN) DspConstants.MEL_MIN else acc
+            val clamped = if (acc < DspConstants.MEL_MIN) DspConstants.MEL_MIN.toDouble() else acc
             mel[m] = ln(clamped).toFloat()
         }
         return mel
